+++ CHANGELOG.md
@@ -41,6 +41,7 @@ en las distintas versiones de la misma.
 * feat: Opción de crear una instancia Docker con datos de demostración
 * feat: Un convenio de FP dual puede ser seguido por responsables de seguimiento opcionales
 * fix: Solucionar un problema al añadir un nuevo responsable o tutor laboral
+* feat: Importar resultados de aprendizaje y criterios de evaluación desde Séneca
 
 4.1.0 (2022-02-13)
 ------------------